Add endpoint GET /countries

+++ src/main/java/com/hotelbooking/repository/AuthorityRepository.java
@@ -4,5 +4,5 @@ import com.hotelbooking.entity.Authority;
 import com.hotelbooking.entity.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AuthorityRepository extends CrudRepository<Authority, Long> {
+public interface AuthorityRepository extends CrudRepository<Authority, Integer> {
 }
