+++ src/main/java/com/hotelbooking/repository/BookingRepository.java
@@ -4,5 +4,5 @@ import com.hotelbooking.entity.Booking;
 import com.hotelbooking.entity.HotelService;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BookingRepository extends CrudRepository<Booking, Long> {
+public interface BookingRepository extends CrudRepository<Booking, Integer> {
 }
