+++ src/main/java/com/hotelbooking/repository/HotelServiceRepository.java
@@ -4,5 +4,5 @@ import com.hotelbooking.entity.HotelService;
 import com.hotelbooking.entity.Room;
 import org.springframework.data.repository.CrudRepository;
 
-public interface HotelServiceRepository extends CrudRepository<HotelService, Long> {
+public interface HotelServiceRepository extends CrudRepository<HotelService, Integer> {
 }
