+++ src/main/java/com/hotelbooking/repository/RoomCategoryRepository.java
@@ -4,5 +4,5 @@ import com.hotelbooking.entity.RoomCategory;
 import com.hotelbooking.entity.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RoomCategoryRepository extends CrudRepository<RoomCategory, Long> {
+public interface RoomCategoryRepository extends CrudRepository<RoomCategory, Integer> {
 }
