+++ src/main/java/com/hotelbooking/repository/RoomRepository.java
@@ -4,5 +4,5 @@ import com.hotelbooking.entity.Room;
 import com.hotelbooking.entity.RoomCategory;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RoomRepository extends CrudRepository<Room, Long> {
+public interface RoomRepository extends CrudRepository<Room, Integer> {
 }
